+++ src/careamics/utils/__init__.py
@@ -5,7 +5,6 @@ __all__ = [
     "denormalize",
     "normalize",
     "get_device",
-    "setup_cudnn_reproducibility",
     "check_array_validity",
     "check_axes_validity",
     "check_tiling_validity",
@@ -18,5 +17,5 @@ __all__ = [
 from .context import cwd
 from .metrics import MetricTracker
 from .normalization import denormalize, normalize
-from .torch_utils import compile_model, get_device, setup_cudnn_reproducibility
+from .torch_utils import compile_model, get_device
 from .validators import check_array_validity, check_axes_validity, check_tiling_validity
